Move all models to data layer
