+++ app/src/main/java/ru/neexol/rtut/core/Constants.kt
@@ -2,9 +2,9 @@ package ru.neexol.rtut.core
 
 object Constants {
 	private const val BASE_URL = "http://192.168.31.177"
-	const val LESSONS_URL = "$BASE_URL/api/schedule"
-	const val NOTES_URL = "$BASE_URL/api/notes"
-	const val MAPS_URL = "$BASE_URL/static/maps"
+	const val LESSONS_URL = "$BASE_URL/api/schedule/"
+	const val NOTES_URL = "$BASE_URL/api/notes/"
+	const val MAPS_URL = "$BASE_URL/static/maps/"
 
 	const val LESSONS_PREFS_NAME = "lessons_prefs"
 }
