--- app/src/main/java/ru/neexol/rtut/data/remote/lessons/models/GroupLessonsResponse.kt
@@ -1,12 +1,0 @@
-package ru.neexol.rtut.data.remote.lessons.models
-
-import ru.neexol.rtut.domain.lessons.models.GroupLessons
-
-@kotlinx.serialization.Serializable
-data class GroupLessonsResponse(
-	val group: String,
-	val checksum: String,
-	val lessons: List<LessonResponse>
-) {
-	fun toGroupLessons() = GroupLessons(group, checksum, lessons.map { it.toLesson() })
-}
