--- app/src/main/java/ru/neexol/rtut/data/remote/lessons/models/LessonResponse.kt
@@ -1,17 +1,0 @@
-package ru.neexol.rtut.data.remote.lessons.models
-
-import ru.neexol.rtut.domain.lessons.models.Lesson
-
-@kotlinx.serialization.Serializable
-data class LessonResponse(
-	val id: String,
-	val name: String,
-	val type: String,
-	val teacher: String,
-	val classroom: String,
-	val day: Int,
-	val number: Int,
-	val weeks: List<Int>
-) {
-	fun toLesson() = Lesson(id, name, type, teacher, classroom, day, number, weeks)
-}
