+++ app/src/main/java/ru/neexol/rtut/domain/lessons/usecases/GetTimes.kt
@@ -1,6 +1,6 @@
 package ru.neexol.rtut.domain.lessons.usecases
 
-import ru.neexol.rtut.data.local.lessons.LessonsLocalDataSource
+import ru.neexol.rtut.data.lessons.local.LessonsLocalDataSource
 import javax.inject.Inject
 
 class GetTimes @Inject constructor(
